pynews_crawler/requirements.txt
attrs==15.2.0
BeautifulSoup==3.2.1
cffi==1.6.0
cryptography==1.3.1
cssselect==0.9.1
enum34==1.1.4
goose-extractor==1.0.25
idna==2.1
ipaddress==1.0.16
jieba==0.38
lxml==3.6.0
nltk==3.2.1
Pillow==3.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pymongo==3.2.2
pyOpenSSL==16.0.0
queuelib==1.4.2
Scrapy==1.0.5
service-identity==16.0.0
six==1.10.0
Twisted==16.1.1
w3lib==1.14.2
zope.interface==4.1.3
